fix: 섹션 컴포넌트로 분리
