+++ src/components/index.ts
@@ -1,1 +1,7 @@
 export * from './ui';
+export { default as Section1 } from './Section1';
+export { default as Section2 } from './Section2';
+export { default as Section3 } from './Section3';
+
+export { default as FeatureBox } from './FeatureBox';
+export { default as Feature } from './Feature';
